wip: use arena for modules to save memory
