--- core_lang/src/parse_tree/expression/match_condition.rs
@@ -1,7 +1,0 @@
-use super::Expression;
-
-#[derive(Debug, Clone)]
-pub(crate) enum MatchCondition {
-    CatchAll,
-    Expression(Box<Expression>),
-}
