--- core_lang/src/semantic_analysis/ast_node/expression/match_branch.rs
@@ -1,10 +1,0 @@
-use super::*;
-use crate::semantic_analysis::ast_node::TypedCodeBlock;
-use either::Either;
-
-#[allow(dead_code)]
-#[derive(Clone, Debug)]
-pub(crate) struct TypedMatchBranch {
-    condition: TypedMatchCondition,
-    result: Either<TypedCodeBlock, TypedExpression>,
-}
