--- core_lang/src/semantic_analysis/ast_node/expression/match_condition.rs
@@ -1,8 +1,0 @@
-use crate::semantic_analysis::TypedExpression;
-
-#[allow(dead_code)]
-#[derive(Clone, Debug)]
-pub(crate) enum TypedMatchCondition {
-    CatchAll,
-    Expression(Box<TypedExpression>),
-}
