+++ sway-core/src/semantic_analysis/ast_node/impl_trait.rs
@@ -20,8 +20,8 @@ use std::collections::{HashMap, HashSet};
 
 pub(crate) fn implementation_of_trait(
     impl_trait: ImplTrait,
-    namespace: &mut Namespace,
-    crate_namespace: Option<&Namespace>,
+    namespace: crate::semantic_analysis::NamespaceRef,
+    crate_namespace: Option<crate::semantic_analysis::NamespaceRef>,
     build_config: &BuildConfig,
     dead_code_graph: &mut ControlFlowGraph,
     dependency_graph: &mut HashMap<String, HashSet<String>>,
@@ -192,8 +192,8 @@ fn type_check_trait_implementation(
     methods: &[FunctionDeclaration],
     trait_name: &Ident,
     type_arguments: &[TypeParameter],
-    namespace: &mut Namespace,
-    crate_namespace: Option<&Namespace>,
+    namespace: crate::semantic_analysis::NamespaceRef,
+    crate_namespace: Option<crate::semantic_analysis::NamespaceRef>,
     _self_type: TypeId,
     build_config: &BuildConfig,
     dead_code_graph: &mut ControlFlowGraph,
