+++ sway-core/src/semantic_analysis/namespace.rs
@@ -44,87 +44,7 @@ impl Namespace {
         self.modules.values()
     }
 
-    /// this function either returns a struct (i.e. custom type), `None`, denoting the type that is
-    /// being looked for is actually a generic, not-yet-resolved type.
-    ///
-    ///
-    /// If a self type is given and anything on this ref chain refers to self, update the chain.
-    pub(crate) fn resolve_type_with_self(
-        &self,
-        ty: TypeInfo,
-        self_type: TypeId,
-    ) -> Result<TypeId, ()> {
-        Ok(match ty {
-            TypeInfo::Custom { ref name } => match self.get_symbol_by_str(name) {
-                Some(TypedDeclaration::StructDeclaration(TypedStructDeclaration {
-                    name,
-                    fields,
-                    ..
-                })) => crate::type_engine::insert_type(TypeInfo::Struct {
-                    name: name.as_str().to_string(),
-                    fields: fields
-                        .iter()
-                        .map(TypedStructField::as_owned_typed_struct_field)
-                        .collect::<Vec<_>>(),
-                }),
-                Some(TypedDeclaration::EnumDeclaration(TypedEnumDeclaration {
-                    name,
-                    variants,
-                    ..
-                })) => crate::type_engine::insert_type(TypeInfo::Enum {
-                    name: name.as_str().to_string(),
-                    variant_types: variants
-                        .iter()
-                        .map(TypedEnumVariant::as_owned_typed_enum_variant)
-                        .collect(),
-                }),
-                Some(TypedDeclaration::GenericTypeForFunctionScope { name, .. }) => {
-                    crate::type_engine::insert_type(TypeInfo::UnknownGeneric {
-                        name: name.as_str().to_string(),
-                    })
-                }
-                _ => return Err(()),
-            },
-            TypeInfo::SelfType => self_type,
-            TypeInfo::Ref(id) => id,
-            o => insert_type(o),
-        })
-    }
 
-    /// Used to resolve a type when there is no known self type. This is needed
-    /// when declaring new self types.
-    pub(crate) fn resolve_type_without_self(&self, ty: &TypeInfo) -> TypeId {
-        let ty = ty.clone();
-        match ty {
-            TypeInfo::Custom { name } => match self.get_symbol_by_str(&name) {
-                Some(TypedDeclaration::StructDeclaration(TypedStructDeclaration {
-                    name,
-                    fields,
-                    ..
-                })) => crate::type_engine::insert_type(TypeInfo::Struct {
-                    name: name.as_str().to_string(),
-                    fields: fields
-                        .iter()
-                        .map(TypedStructField::as_owned_typed_struct_field)
-                        .collect::<Vec<_>>(),
-                }),
-                Some(TypedDeclaration::EnumDeclaration(TypedEnumDeclaration {
-                    name,
-                    variants,
-                    ..
-                })) => crate::type_engine::insert_type(TypeInfo::Enum {
-                    name: name.as_str().to_string(),
-                    variant_types: variants
-                        .iter()
-                        .map(TypedEnumVariant::as_owned_typed_enum_variant)
-                        .collect(),
-                }),
-                _ => crate::type_engine::insert_type(TypeInfo::Unknown),
-            },
-            TypeInfo::Ref(id) => id,
-            o => insert_type(o),
-        }
-    }
 
     pub(crate) fn insert(&mut self, name: Ident, item: TypedDeclaration) -> CompileResult<()> {
         let mut warnings = vec![];
@@ -153,172 +73,6 @@ impl Namespace {
         ok((), warnings, errors)
     }
 
-    // TODO(static span) remove this and switch to spans when we have arena spans
-    pub(crate) fn get_symbol_by_str(&self, symbol: &str) -> Option<TypedDeclaration> {
-        let empty = vec![];
-        let path = self
-            .use_synonyms
-            .iter()
-            .find_map(|(name, value)| {
-                if name.as_str() == symbol {
-                    Some(value)
-                } else {
-                    None
-                }
-            })
-            .unwrap_or(&empty);
-        self.get_name_from_path_str(path, symbol).value
-    }
-
-    pub(crate) fn get_symbol(&self, symbol: &Ident) -> CompileResult<TypedDeclaration> {
-        let empty = vec![];
-        let path = self.use_synonyms.get(symbol).unwrap_or(&empty);
-        let true_symbol = self
-            .use_aliases
-            .get(&symbol.as_str().to_string())
-            .unwrap_or(symbol);
-        self.get_name_from_path(path, true_symbol)
-    }
-
-    /// Used for calls that look like this:
-    /// `foo::bar::function`
-    /// where `foo` and `bar` are the prefixes
-    /// and `function` is the suffix
-    pub(crate) fn get_call_path(&self, symbol: &CallPath) -> CompileResult<TypedDeclaration> {
-        let path = if symbol.prefixes.is_empty() {
-            self.use_synonyms
-                .get(&symbol.suffix)
-                .unwrap_or(&symbol.prefixes)
-        } else {
-            &symbol.prefixes
-        };
-        self.get_name_from_path(path, &symbol.suffix)
-            .map(|decl| decl.clone())
-    }
-
-    fn get_name_from_path(&self, path: &[Ident], name: &Ident) -> CompileResult<TypedDeclaration> {
-        let mut warnings = vec![];
-        let mut errors = vec![];
-        let module = check!(
-            self.find_module_relative(path),
-            return err(warnings, errors),
-            warnings,
-            errors
-        );
-        match read_module(|module| module.symbols.get(name).cloned(), module) {
-            Some(decl) => ok(decl, warnings, errors),
-            None => {
-                errors.push(CompileError::SymbolNotFound {
-                    name: name.as_str().to_string(),
-                    span: name.span().clone(),
-                });
-                err(warnings, errors)
-            }
-        }
-    }
-
-    // TODO(static span) remove this when typeinfo uses spans
-    fn get_name_from_path_str(
-        &self,
-        path: &[Ident],
-        name: &str,
-    ) -> CompileResult<TypedDeclaration> {
-        let mut warnings = vec![];
-        let mut errors = vec![];
-        let module = check!(
-            self.find_module_relative(path),
-            return err(warnings, errors),
-            warnings,
-            errors
-        );
-
-        match read_module(
-            |module| {
-                module
-                    .symbols
-                    .iter()
-                    .find_map(|(item, other)| {
-                        if item.as_str() == name {
-                            Some(other)
-                        } else {
-                            None
-                        }
-                    })
-                    .cloned()
-            },
-            module,
-        ) {
-            Some(decl) => ok(decl, warnings, errors),
-            None => {
-                let span = match path.get(0) {
-                    Some(ident) => ident.span().clone(),
-                    None => {
-                        errors.push(CompileError::Internal("Unable to construct span. This is a temporary error and will be fixed in a future release. )", Span { span: pest::Span::new(" ".into(), 0, 0).unwrap(),
-                                path: None
-                            }));
-                        Span {
-                            span: pest::Span::new(" ".into(), 0, 0).unwrap(),
-                            path: None,
-                        }
-                    }
-                };
-                errors.push(CompileError::SymbolNotFound {
-                    name: name.to_string(),
-                    span,
-                });
-                err(warnings, errors)
-            }
-        }
-    }
-
-    pub(crate) fn find_module_relative(&self, path: &[Ident]) -> CompileResult<NamespaceRef> {
-        debug_assert!(!path.is_empty());
-        let mut namespace = self;
-        let mut errors = vec![];
-        let warnings = vec![];
-        let ix = self.modules.get(path[0].as_str());
-        let mut ix: NamespaceRef = match ix {
-            Some(ix) => *ix,
-            None => {
-                errors.push(CompileError::ModuleNotFound {
-                    span: path.iter().fold(path[0].span().clone(), |acc, this_one| {
-                        join_spans(acc, this_one.span().clone())
-                    }),
-                    name: path
-                        .iter()
-                        .map(|x| x.as_str())
-                        .collect::<Vec<_>>()
-                        .join("::"),
-                });
-                return err(warnings, errors);
-            }
-        };
-        for ident in path {
-            match read_module(
-                |namespace| namespace.modules.get(ident.as_str()).cloned(),
-                ix,
-            ) {
-                Some(ns_ix) => {
-                    ix = ns_ix;
-                }
-                _ => {
-                    errors.push(CompileError::ModuleNotFound {
-                        span: path.iter().fold(path[0].span().clone(), |acc, this_one| {
-                            join_spans(acc, this_one.span().clone())
-                        }),
-                        name: path
-                            .iter()
-                            .map(|x| x.as_str())
-                            .collect::<Vec<_>>()
-                            .join("::"),
-                    });
-                    return err(warnings, errors);
-                }
-            };
-        }
-
-        ok(ix, warnings, errors)
-    }
 
     pub(crate) fn insert_trait_implementation(
         &mut self,
@@ -353,31 +107,18 @@ impl Namespace {
         ok((), warnings, errors)
     }
 
-    pub fn insert_module(&mut self, module_name: String, module_contents: Namespace) {
-        let ix = {
-            let mut write_lock = MODULES.write().expect("poisoned lock");
-            write_lock.insert(module_contents)
-        };
+    pub fn insert_module(&mut self, module_name: String, ix: NamespaceRef) {
         self.modules.insert(module_name, ix);
     }
 
-    pub fn insert_dependency_module(&mut self, module_name: String, module_contents: Namespace) {
-        self.insert_module(module_name, module_contents)
+    pub fn insert_dependency_module(&mut self, module_name: String, ix: NamespaceRef) {
+        self.insert_module(module_name, ix)
     }
 
-    pub(crate) fn find_enum(&self, enum_name: &Ident) -> Option<TypedEnumDeclaration> {
-        match self.get_symbol(enum_name) {
-            CompileResult {
-                value: Some(TypedDeclaration::EnumDeclaration(inner)),
-                ..
-            } => Some(inner.clone()),
-            _ => None,
-        }
-    }
     /// Returns a tuple where the first element is the [ResolvedType] of the actual expression,
     /// and the second is the [ResolvedType] of its parent, for control-flow analysis.
     pub(crate) fn find_subfield_type(
-        &mut self,
+        &self,
         subfield_exp: &[Ident],
     ) -> CompileResult<(TypeId, TypeId)> {
         let mut warnings = vec![];
@@ -520,196 +261,6 @@ impl Namespace {
                     actually: a.friendly_type_str(),
                 }],
             ),
-        }
-    }
-
-    /// Given a path to a module, create synonyms to every symbol in that module.
-    /// This is used when an import path contains an asterisk.
-    pub(crate) fn star_import(
-        &mut self,
-        from_module: Option<&Namespace>,
-        path: Vec<Ident>,
-    ) -> CompileResult<()> {
-        let mut warnings = vec![];
-        let mut errors = vec![];
-        let namespace = {
-            let base_namespace = match from_module {
-                Some(base_namespace) => base_namespace,
-                None => self,
-            };
-            check!(
-                base_namespace.find_module_relative(&path),
-                return err(warnings, errors),
-                warnings,
-                errors
-            )
-        };
-        let (symbols, implemented_traits) = read_module(
-            |namespace| {
-                let symbols = namespace
-                    .symbols
-                    .iter()
-                    .filter_map(|(symbol, decl)| {
-                        if decl.visibility() == Visibility::Public {
-                            Some(symbol.clone())
-                        } else {
-                            None
-                        }
-                    })
-                    .collect::<Vec<_>>();
-                (symbols, namespace.implemented_traits.clone())
-            },
-            namespace,
-        );
-        self.implemented_traits
-            .extend(&mut implemented_traits.into_iter());
-        for symbol in symbols {
-            self.use_synonyms.insert(symbol, path.clone());
-        }
-        ok((), warnings, errors)
-    }
-
-    /// Pull a single item from a module and import it into this namespace.
-    pub(crate) fn item_import(
-        &mut self,
-        from_namespace: Option<&Namespace>,
-        path: Vec<Ident>,
-        item: &Ident,
-        alias: Option<Ident>,
-    ) -> CompileResult<()> {
-        let mut warnings = vec![];
-        let mut errors = vec![];
-        let base_namespace = match from_namespace {
-            Some(base_namespace) => base_namespace,
-            None => self,
-        };
-        let namespace = check!(
-            base_namespace.find_module_relative(&path),
-            return err(warnings, errors),
-            warnings,
-            errors
-        );
-        let mut impls_to_insert = vec![];
-
-        match read_module(|namespace| namespace.symbols.get(item).cloned(), namespace) {
-            Some(decl) => {
-                //  if this is an enum or struct, import its implementations
-                if decl.visibility() != Visibility::Public {
-                    errors.push(CompileError::ImportPrivateSymbol {
-                        name: item.as_str().to_string(),
-                        span: item.span().clone(),
-                    });
-                }
-                let a = decl.return_type().value;
-                let mut res = read_module(
-                    move |namespace| {
-                        namespace
-                            .implemented_traits
-                            .iter()
-                            .filter(|((_trait_name, type_info), _impl)| {
-                                a.map(look_up_type_id).as_ref() == Some(type_info)
-                            })
-                            .fold(Vec::new(), |mut acc, (a, b)| {
-                                acc.push((a.clone(), b.to_vec()));
-                                acc
-                            })
-                    },
-                    namespace,
-                );
-                impls_to_insert.append(&mut res);
-                // no matter what, import it this way though.
-                match alias {
-                    Some(alias) => {
-                        self.use_synonyms.insert(alias.clone(), path);
-                        self.use_aliases
-                            .insert(alias.as_str().to_string(), item.clone());
-                    }
-                    None => {
-                        self.use_synonyms.insert(item.clone(), path);
-                    }
-                };
-            }
-            None => {
-                errors.push(CompileError::SymbolNotFound {
-                    name: item.as_str().to_string(),
-                    span: item.span().clone(),
-                });
-                return err(warnings, errors);
-            }
-        };
-
-        impls_to_insert.into_iter().for_each(|(a, b)| {
-            self.implemented_traits.insert(a, b);
-        });
-
-        ok((), warnings, errors)
-    }
-
-    /// Given a method and a type (plus a `self_type` to potentially resolve it), find that
-    /// method in the namespace. Requires `args_buf` because of some special casing for the
-    /// standard library where we pull the type from the arguments buffer.
-    ///
-    /// This function will generate a missing method error if the method is not found.
-    pub(crate) fn find_method_for_type(
-        &self,
-        r#type: TypeId,
-        method_name: &Ident,
-        method_path: &[Ident],
-        from_module: Option<&Namespace>,
-        self_type: TypeId,
-        args_buf: &VecDeque<TypedExpression>,
-    ) -> CompileResult<TypedFunctionDeclaration> {
-        let mut warnings = vec![];
-        let mut errors = vec![];
-        let base_module = match from_module {
-            Some(base_module) => base_module,
-            None => self,
-        };
-        let namespace = check!(
-            base_module.find_module_relative(method_path),
-            return err(warnings, errors),
-            warnings,
-            errors
-        );
-
-        // This is a hack and I don't think it should be used.  We check the local namespace first,
-        // but if nothing turns up then we try the namespace where the type itself is declared.
-        let r#type = read_module(
-            |namespace| namespace.resolve_type_with_self(look_up_type_id(r#type), self_type),
-            namespace,
-        )
-        .unwrap_or_else(|_| {
-            errors.push(CompileError::UnknownType {
-                span: method_name.span().clone(),
-            });
-            insert_type(TypeInfo::ErrorRecovery)
-        });
-        let local_methods = self.get_methods_for_type(r#type);
-        let mut ns_methods = read_module(
-            |namespace| namespace.get_methods_for_type(r#type),
-            namespace,
-        );
-
-        let mut methods = local_methods;
-        methods.append(&mut ns_methods);
-
-        match methods
-            .into_iter()
-            .find(|TypedFunctionDeclaration { name, .. }| name == method_name)
-        {
-            Some(o) => ok(o, warnings, errors),
-            None => {
-                if args_buf.get(0).map(|x| look_up_type_id(x.return_type))
-                    != Some(TypeInfo::ErrorRecovery)
-                {
-                    errors.push(CompileError::MethodNotFound {
-                        method_name: method_name.as_str().to_string(),
-                        type_name: r#type.friendly_type_str(),
-                        span: method_name.span().clone(),
-                    });
-                }
-                err(warnings, errors)
-            }
         }
     }
 }
